Wrap view monthly availability behind suspense boundary
